refactor: caching v3 (#100)
refactor: rewrite caching and reorganize state
